:sparkles: Déplacement du code fillVariable en treatVariable
